feat: :sparkles: Added max_stack_size and gc_threshold to vmoptions and added ModuleResolver back in
